backend/app/pyproject.toml
[project]
name = "app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "openai>=1.78.1",
    "pymupdf>=1.25.5",
    "python-dotenv>=1.1.0",
]
